requirements.txt
albumentations==0.4.5
backcall==0.2.0
certifi==2020.4.5.2
cycler==0.10.0
decorator==4.4.2
efficientnet-pytorch==0.6.3
future==0.18.2
imageio==2.8.0
imgaug==0.2.6
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.1
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.2
munch==2.5.0
networkx==2.4
numpy==1.18.1
olefile==0.46
opencv-python-headless==4.2.0.34
pandas==1.0.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pretrainedmodels==0.7.4
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2020.6.3
torch==1.5.1
tqdm==4.46.1
traitlets==4.3.3
wcwidth==0.2.4
